Corrected sysPath and enabled runtime:inline
Brunch completely failed to build when I first added this plugin. This
was because the author for got to define sysPath.

After I fixed this issue, another error appeared informing me that the
vendor dependency did not exist. I commented out the line, which
allowed brunch to build my application. However this caused a
javascript error as the vendor dependency was important to
IcedCoffeeScript. I tried downloading a similarly named file from the
ICS website and placing it in the vendor directory, but the module was
not detected by the brunch build system.

I eventually decided to change the compile runtime option to 'inline'.
This removed the need for the dependency completely. It does, however,
create unnecessary code duplication throughout the application as the
runtime functions are redefined in every module they are used. I might
come back to this later and try and get the js module to work with
brunch, but I've wasted enough time on this already.

I've left sysPath defined despite the fact it's not really required
anymore. Feel free to remove it.